+++ fastMedicineServer/src/main/java/com/nicholas/fastmedicineserver/repository/FeedbackRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nicholas.fastmedicineserver.entity.Feedback;
 
-public interface FeedbackRepository extends JpaRepository<Feedback, Long>
+public interface FeedbackRepository extends JpaRepository<Feedback, Integer>
 {
 
 }
