+++ fastMedicineServer/src/main/java/com/nicholas/fastmedicineserver/repository/ProductDetailRepository.java
@@ -10,11 +10,11 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.nicholas.fastmedicineserver.entity.ProductDetail;
 
-public interface ProductDetailRepository extends JpaRepository<ProductDetail, Long>
+public interface ProductDetailRepository extends JpaRepository<ProductDetail, Integer>
 {
-	ProductDetail findById(Long id);
+	ProductDetail findById(Integer id);
 	
-	List<ProductDetail> findByCategoryId(Long id);
+	List<ProductDetail> findByCategoryId(Integer id);
 	
 	
 	@Query(nativeQuery = true, value ="SELECT * from t_productdetail WHERE  product_key LIKE ?1")
