+++ fastMedicineServer/src/main/java/com/nicholas/fastmedicineserver/repository/UserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nicholas.fastmedicineserver.entity.UserInfo;
 
-public interface UserRepository extends JpaRepository<UserInfo, Long>
+public interface UserRepository extends JpaRepository<UserInfo, Integer>
 {
 		UserInfo findByPhoneNum(String num);
